Combined my zoo website to my web solutions website. Figured I would combine the websites in order to open up a slot for another website.
I customized the CSS, HTML, Javascript, the navbar, the initial code.
Literally, everything was done by other than the navigation bar. The
navigation bar is completely changed as well as it's position on the
website. It's position was fixed on the screen. I made it stationary
transparent with a light border around the links to navigate throughout
the page. It's less interactive then websolutions.today however, this
website is strictly geared for cellphone users. I advertise my services
on LinkedIn, Facebook, and occasionally forums. From that I was able to
realize that most of my incoming users were using mobile cellphones
using a third-party app that watched my user traffic and told me where
people were mostly coming from. With the collected data I created this
website in order to test out which website gained more traction. Without
a doubt this website and my websolutions.today website get the most user
traffic.  The reason why is I think it's because of the way I chose to
design it. It's strictly for MOBILE USERS. We'll see what comes in the
future... stay tuned!:)